LPD-15137 Convert App to typescript and fetch currentUser only one time
