+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/App.tsx
@@ -1,0 +1,368 @@
+import AlertFeed from 'shared/components/AlertFeed';
+import BundleRouter from './route-middleware/BundleRouter';
+import ChannelProvider from 'shared/context/channel';
+import client from 'shared/apollo/client';
+import ErrorPage from 'shared/pages/ErrorPage';
+import Loading from 'shared/components/Loading';
+import ModalRenderer from 'shared/components/ModalRenderer';
+import pathToRegexp from 'path-to-regexp';
+import React, {lazy, Suspense, useEffect, useState} from 'react';
+import RouteNotFound from 'shared/components/RouteNotFound';
+import store from 'shared/store';
+import UnassignedSegmentsProvider from 'shared/context/unassignedSegments';
+import {ApolloProvider} from '@apollo/react-components';
+import {ApolloProvider as ApolloProviderHooks} from '@apollo/react-hooks';
+import {ClayIconSpriteContext} from '@clayui/icon';
+import {ClayLinkContext} from '@clayui/link';
+import {ClayTooltipProvider} from '@clayui/tooltip';
+import {close, modalTypes, open} from 'shared/actions/modals';
+import {
+	Link,
+	matchPath,
+	Route,
+	BrowserRouter as Router,
+	Switch,
+	useLocation
+} from 'react-router-dom';
+import {OnboardingContext} from 'shared/context/onboarding';
+import {PROD_MODE, spritemap} from 'shared/util/constants';
+import {Provider, useSelector} from 'react-redux';
+import {Routes} from 'shared/util/router';
+import {saveState} from 'shared/store/local-storage';
+import {setBackURL} from 'shared/actions/settings';
+import {throttle} from 'lodash';
+import {useFetchCurrentUser} from 'shared/hooks/useCurrentUser';
+
+// Workspaces
+
+const AddWorkspace = lazy(
+	() =>
+		import(
+			/* webpackChunkName: "AddWorkspace" */ './shared/pages/AddWorkspace'
+		)
+);
+const SelectWorkspaceAccount = lazy(
+	() =>
+		import(
+			/* webpackChunkName: "SelectWorkspaceAccount" */ './shared/pages/SelectWorkspaceAccount'
+		)
+);
+const Workspaces = lazy(
+	() =>
+		import(/* webpackChunkName: "Workspaces" */ './shared/pages/Workspaces')
+);
+
+// WorkspaceLayer
+const WorkspaceLayer = lazy(
+	() =>
+		import(
+			/* webpackChunkName: "WorkspaceLayer" */ './shared/components/WorkspaceLayer'
+		)
+);
+
+// Other
+
+const OAuthReceive = lazy(
+	() =>
+		import(
+			/* webpackChunkName: "OAuthReceive" */ './settings/pages/OAuthReceive'
+		)
+);
+
+const SETTINGS_PATH_REGEX = pathToRegexp(Routes.SETTINGS, null, {end: false});
+
+const RoutesContainer = ({children}) => {
+	const location = useLocation();
+
+	const matchingPath = matchPath<any>(location.pathname, {
+		path: Routes.WORKSPACE_WITH_ID
+	});
+
+	const groupId = matchingPath?.params.groupId ?? '0';
+
+	const {data: currentUser, loading} = useFetchCurrentUser(groupId);
+
+	const project = useSelector<any, any>(state =>
+		state.getIn(['projects', groupId, 'data'])
+	);
+
+	const pendoFn =
+		pendo && pendo.isReady && pendo.isReady()
+			? pendo.identify
+			: pendo.initialize;
+
+	useEffect(() => {
+		const {
+			document: {referrer, title},
+			location: {href, pathname, search}
+		} = window;
+
+		analytics.page(
+			{
+				path: pathname,
+				referrer,
+				search,
+				title,
+				url: href
+			},
+			{ip: '0'}
+		);
+
+		if (!SETTINGS_PATH_REGEX.test(pathname)) {
+			store.dispatch(setBackURL(`${pathname}${search}`));
+		}
+	}, [location]);
+
+	useEffect(() => {
+		let account;
+		let visitor;
+
+		if (currentUser) {
+			// We use userId instead of id because userId persists across workspaces.
+			const {emailAddress, name, roleName, userId} = currentUser;
+
+			visitor = {
+				emailAddress,
+				id: userId,
+				name,
+				roleName
+			};
+
+			analytics && analytics.identify(userId, null, {ip: '0'});
+		}
+
+		if (project) {
+			const {
+				corpProjectName,
+				corpProjectUuid,
+				faroSubscription: faroSubscriptionIMap,
+				groupId,
+				name: workspaceName,
+				ownerEmailAddress: workspaceOwnerEmailAddress,
+				serverLocation
+			} = project;
+
+			const subscriptionName = faroSubscriptionIMap.get('name');
+
+			account = {
+				corpProjectNameId: `${corpProjectName}: ${corpProjectUuid}`,
+				id: groupId,
+				serverLocation,
+				subscriptionName,
+				workspaceName,
+				workspaceOwnerEmailAddress
+			};
+
+			analytics &&
+				analytics.group(
+					groupId,
+					{
+						groupId,
+						serverLocation,
+						subscriptionName,
+						workspaceName
+					},
+					{ip: '0'}
+				);
+		}
+
+		if (account || visitor) {
+			pendoFn &&
+				pendoFn({
+					account,
+					visitor
+				});
+		}
+	}, [location, currentUser, project]);
+
+	if (loading) {
+		return <Loading />;
+	}
+
+	if (location?.state?.notFoundError) {
+		return <ErrorPage />;
+	}
+
+	return children;
+};
+
+const App = () => {
+	const [onboardingTriggered, setOnboardingTriggered] = useState(false);
+
+	useEffect(() => {
+		store.subscribe(throttle(() => saveState(store.getState()), 1000));
+	}, []);
+
+	const handleUserConfirmation = (message, callback) => {
+		store.dispatch(
+			open(modalTypes.CONFIRMATION_MODAL, {
+				cancelMessage: Liferay.Language.get('stay-on-page'),
+				message,
+				modalVariant: 'modal-warning',
+				onClose: () => {
+					callback(false);
+
+					store.dispatch(close());
+				},
+				onSubmit: () => {
+					callback(true);
+				},
+				submitButtonDisplay: 'warning',
+				submitMessage: Liferay.Language.get('leave-page'),
+				title: Liferay.Language.get('unsaved-changes'),
+				titleIcon: 'warning-full'
+			})
+		);
+	};
+
+	return (
+		<ApolloProvider client={client}>
+			<ApolloProviderHooks client={client}>
+				<Provider store={store}>
+					<ClayIconSpriteContext.Provider value={spritemap}>
+						<ClayLinkContext.Provider
+							value={({children, href, ...otherProps}: any) => {
+								if (href?.startsWith('http')) {
+									return (
+										<a {...otherProps} href={href}>
+											{children}
+										</a>
+									);
+								}
+
+								return (
+									<Link {...otherProps} to={href}>
+										{children}
+									</Link>
+								);
+							}}
+						>
+							<UnassignedSegmentsProvider>
+								<OnboardingContext.Provider
+									value={{
+										onboardingTriggered,
+										setOnboardingTriggered: () =>
+											setOnboardingTriggered(true)
+									}}
+								>
+									<ChannelProvider>
+										<ClayTooltipProvider>
+											<div>
+												<Router
+													getUserConfirmation={
+														handleUserConfirmation
+													}
+												>
+													<RoutesContainer>
+														<AlertFeed />
+
+														<ModalRenderer />
+
+														<Suspense
+															fallback={
+																<Loading />
+															}
+														>
+															<Switch>
+																<BundleRouter
+																	data={
+																		Workspaces
+																	}
+																	exact
+																	path={
+																		Routes.BASE
+																	}
+																/>
+
+																<BundleRouter
+																	data={
+																		Workspaces
+																	}
+																	exact
+																	path={
+																		Routes.WORKSPACES
+																	}
+																/>
+
+																<BundleRouter
+																	data={
+																		SelectWorkspaceAccount
+																	}
+																	exact
+																	path={
+																		Routes.WORKSPACE_ADD
+																	}
+																/>
+
+																{!PROD_MODE && (
+																	<BundleRouter
+																		data={
+																			AddWorkspace
+																		}
+																		exact
+																		path={
+																			Routes.WORKSPACE_ADD_TRIAL
+																		}
+																	/>
+																)}
+
+																<BundleRouter
+																	data={
+																		AddWorkspace
+																	}
+																	exact
+																	path={
+																		Routes.WORKSPACE_ADD_WITH_CORP_PROJECT_UUID
+																	}
+																/>
+
+																<BundleRouter
+																	data={
+																		SelectWorkspaceAccount
+																	}
+																	exact
+																	path={
+																		Routes.WORKSPACE_SELECT_ACCOUNT
+																	}
+																/>
+
+																<BundleRouter
+																	data={
+																		OAuthReceive
+																	}
+																	exact
+																	path={
+																		Routes.OAUTH_RECEIVE
+																	}
+																/>
+
+																<Route
+																	component={
+																		Loading
+																	}
+																	path={
+																		Routes.LOADING
+																	}
+																/>
+
+																<WorkspaceLayer />
+
+																<RouteNotFound />
+															</Switch>
+														</Suspense>
+													</RoutesContainer>
+												</Router>
+											</div>
+										</ClayTooltipProvider>
+									</ChannelProvider>
+								</OnboardingContext.Provider>
+							</UnassignedSegmentsProvider>
+						</ClayLinkContext.Provider>
+					</ClayIconSpriteContext.Provider>
+				</Provider>
+			</ApolloProviderHooks>
+		</ApolloProvider>
+	);
+};
+
+export default App;
